+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `User` model to improve readability of user repo results
+- `last_sync_at` date to `Watcher` model
 
 ## [0.12.0] - 2023-12-09